refactor: move into a class
